requirements.txt
Django==2.0.2
django-ckeditor==5.6.1
django-js-asset==1.1.0
mysqlclient==1.3.13
Pillow==5.2.0
pycryptodome==3.6.6
pytz==2018.5
